feat: add SuspendMythware module; fix: fix KillMythware module

+++ Modules/MYTHWARE/KillMythware.cs
@@ -1,21 +1,44 @@
-﻿using System.Diagnostics;
+﻿using System.ComponentModel;
+using System.Diagnostics;
 using Vape_for_Windows.Common.Module;
+using Vape_for_Windows.ui;
 
 namespace Vape_for_Windows.Modules.MYTHWARE
 {
     internal class KillMythware : Module
     {
+        private Boolean _killed = false;
+
         public KillMythware() : base("KillMythware", Category.MYTHWARE)
         {
         }
 
         public override void OnEnable()
         {
-            Process[] process = Process.GetProcessesByName("StudentMain.exe");
-            foreach (Process p in process)
+            foreach (Process p in Process.GetProcesses())
             {
-                Debug.WriteLine(p.ExitCode);
-                p.Kill();
+                if (p.ProcessName.Contains("StudentMain"))
+                {
+                    try
+                    {
+                        p.Kill();
+                        p.WaitForExit();
+                        _killed = true;
+                        new NotificationWindow().send("Success", "Module KillMythware excuted successfully!", 0, 5);
+                    }
+                    catch (Win32Exception e)
+                    {
+                        new NotificationWindow().send("Failed", e.Message.ToString(), 0, 5);
+                    }
+                    catch (InvalidOperationException e)
+                    {
+                        new NotificationWindow().send("Failed", e.Message.ToString(), 0, 5);
+                    }
+                }
+            }
+            if (!_killed)
+            {
+                new NotificationWindow().send("Failed", "Process StudentMain.exe no found! ", 0, 5);
             }
         }
     }
